extracting testing code
